refactor: update import paths for moved kmeans1d deps folder
Completes the previous move: repoints pyproject.toml package-data and
the three call sites at coreai_opt.deps._kmeans1d instead of the old
coreai_opt._utils._kmeans1d location. Also anchors the .gitignore
deps/ rule to the repo root (/deps/) so it stops shadowing the new
src/coreai_opt/deps/ package.
