+++ src/coreai_opt/deps/_kmeans1d/__init__.py
@@ -3,6 +3,6 @@
 # Use of this source code is governed by a BSD-3-Clause license that can
 # be found in the LICENSE file or at https://opensource.org/licenses/BSD-3-Clause
 
-from coreai_opt._utils._kmeans1d.core import Clustered, cluster
+from coreai_opt.deps._kmeans1d.core import Clustered, cluster
 
 __all__ = ["Clustered", "cluster"]
